A basic unit test for the provision method
This will very probably be thrown away - and if not will be very much modified,
but I wanted a test round the provision method before we start refactoring. I
went down many rabbit holes before eventually making one small change (previous
commit) in order to get this test to pass.

Some refactoring to do even in the test - not sure we need everything we've
mocked - but that's a task for another day.

cc /@mikepea