Tiny comment change
This hack upsets me, but the alternative appears to be writing our own symbolize_keys method, which has to be recursive but only for nested Hashes and nested Arrays. Complex piece of code for us to maintain vs dirty hack but using an established library? Hmm. Leaving as is for now after some playing with it.