Internalise two of the VSE examples
Rather than linking out to personal gists. Because Sneha may not keep them
forever, and they should be version controlled here.

The NAT example is missing because GitHub currenlty returns:
```
Hm.

This gist can't be viewed at the moment. Please send a message to Support and we'll have a look as soon as we can.
```

@snehaso, do you have another copy of this?