[#63195228] adjust integration_test to ensure skyscape and carrenza have same coverage
Due to the Carrenza templateVapp having 2CPU + 4096MB memory, the integration tests were not calling through to put_memory or put_cpu.

This does mean that these API requests were not being made, and we should really make them in an integration test.

Upped to 8192MB memory and 4 cpus, to ensure that they are covered.

#