[#62448006] display examples/ dir in vcloud-launch help
The configuration file examples are specific to the version
of vcloud-launch in use. It does not make too much sense to
point users to the github docs, since these are potentially
incompatible.

This commit displays the examples directory in direct relation
to the vcloud-launch executable called - so should guarantee they
are useful.