Call Differ directly instead of via diff
We want to use the Differ class instead of the diff method so this small change means no code calls it.
We had to comment out the unit test because they are tied very closely to the implementation - they expect the object to call certain methods etc, which is exactly what we are changing. We are using the integration test to check our refactor - that is green after this change.