No need for jenkins_integration_tests
There is now no Rakefile and no integration tests so no need for the jenkins_integration_tests script. I have also removed the integration test jobs in Jenkins.

This also removes `bundle exec rake` from `jenkins.sh` for similar reasons. However, the the script will be required when the step is made to meta-gem so no need to remove it here.