Update for recent changer to launcher related files
In the past the vcloud-tools included vcloud/launch.rb and a few other
files.  These recently moved to their own gem vcloud-launcher.

Since that change though this repo needs to be updated to include the
dependency and to account for some changes from vcloud/launch to
vcloud/launcher etc.  This commit does that.