We only require Core, Fog and Version
The gemspec will take care of installing all the gems if a user does `gem install vcloud-tools`.  However, if they want to build a script then it is better for them to explicitly require the particular tools they want to use as and when they want to use them as this will make it clearer what is going on, and will allow them to build something that only uses the tools they require.

My thoughts on this inspired by what rspec does, and this specific discussion: https://github.com/rspec/rspec/commit/f10bedd498dff2961d4be9cce793fa152c4a0bbe