Move to Tools namespace
Since this repo exists as a meta-gem, purely for installing all the gems in one go, there is no need to follow the convention of having a `version.rb` that is required by `tools.rb` and then required by the gemspec - since there are no other files to be required or not it doesn't make a difference and could be hard-coded in `tools.rb` or even the gemspec.

However, I've decided to use the convention here as well because that makes it consistent with the other tools. Most of the people working on these tools will not be purely Ruby devs by trade, (e.g. may be web ops, developers in other languages, etc) so it's especially important to follow the principle of least surprise for people working on these tools.