Use website instead of README
Remove everything from the README that is duplicated on the website.

I noticed that there was a credentials section in here that describes usage
prior to `vcloud-login`. The `/usage` section of the website has the correct
details. Then I also noticed that the installation instructions and list of
repos and their descriptions is duplicated on the website.

We should maintain one or the other. Otherwise one is bound to be
out-of-date. This also makes the website link more prominent and less likely
to be skipped over.