Add integration test instructions to site
This commit copies the information that is currently in, e.g. https://github.com/gds-operations/vcloud-core/blob/master/spec/integration/README.md, to the website. This is because there is a lot of detail involved in the setup of intergraion tests and the only bits that vary between projects are the parameters.

I have modified the text slightly but it is in substance the same. One thing I have not changed is the now incorrect instruction as to how to run the tests, which is what prompted this move - to avoid having to update each repo separately.

If this is merged, I will follow up by removing the duplicate information from the READMEs.