Add blog posts on About page
We've written loads of great stuff about vCloud Tools and we'll probably write more; it would be good to includelinks to them on our website.

I added an About page rather than lengthening the index page because I think the index page is good as it is and this should just be available if people want more info.