Amend integration test command
The command mentioned in the README to run the integration tests is wrong,
since `integration` is not a standalone Gem but rather defined in the
Rakefile. As such, it needs to be called through `rake`.

When running the command as-is, the only output is that `integration` could
not be found with a suggestion to install Gems from the Gemfile using
Bundler.

Updates the command to reflect this.